# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: GigShield API
  description: Parametric Income Protection for Q-Commerce Delivery Partners
  version: 2.1.0
paths:
...
  /api/policies/create:
    post:
      tags:
      - Policies
      summary: Create Policy
      description: Create a new insurance policy with dynamic premium.
      operationId: create_policy_api_policies_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/policies/me:
    get:
      tags:
      - Policies
      summary: Get My Policies
      description: Get all policies for the authenticated worker.
      operationId: get_my_policies_api_policies_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/policies/{policy_id}/renew:
    put:
      tags:
      - Policies
      summary: Renew Policy
      description: Renew policy for another week.
      operationId: renew_policy_api_policies__policy_id__renew_put
      security:
      - HTTPBearer: []
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: integer
          title: Policy Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/policies/{policy_id}/pause:
    put:
      tags:
      - Policies
      summary: Pause Policy
      description: Pause coverage — no lock-in, no penalty.
      operationId: pause_policy_api_policies__policy_id__pause_put
      security:
      - HTTPBearer: []
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: integer
          title: Policy Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/policies/{policy_id}/activate:
    put:
      tags:
      - Policies
      summary: Activate Policy
      description: Reactivate a paused policy.
      operationId: activate_policy_api_policies__policy_id__activate_put
      security:
      - HTTPBearer: []
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: integer
          title: Policy Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/policies/all:
    get:
      tags:
      - Policies
      summary: List All Policies
      description: 'Admin: List all policies.'
      operationId: list_all_policies_api_policies_all_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /api/claims/me:
    get:
      tags:
      - Claims
      summary: Get My Claims
      description: Get all claims for the authenticated worker with trigger details.
      operationId: get_my_claims_api_claims_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/claims/worker/{worker_id}:
    get:
      tags:
      - Claims
      summary: Get Worker Claims
      description: Admin-only worker claims lookup.
      operationId: get_worker_claims_api_claims_worker__worker_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: worker_id
        in: path
        required: true
        schema:
          type: integer
          title: Worker Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/claims/{claim_id}:
    get:
      tags:
      - Claims
      summary: Get Claim Detail
      description: Get detailed claim information with fraud analysis.
      operationId: get_claim_detail_api_claims__claim_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: claim_id
        in: path
        required: true
        schema:
          type: integer
          title: Claim Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/claims/{claim_id}/approve:
    put:
      tags:
      - Claims
      summary: Approve
      description: 'Admin: Approve a pending claim.'
      operationId: approve_api_claims__claim_id__approve_put
      security:
      - HTTPBearer: []
      parameters:
      - name: claim_id
        in: path
        required: true
        schema:
          type: integer
          title: Claim Id
      - name: notes
        in: query
        required: false
        schema:
          type: string
          title: Notes
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/claims/{claim_id}/reject:
    put:
      tags:
      - Claims
      summary: Reject
      description: 'Admin: Reject a pending claim.'
      operationId: reject_api_claims__claim_id__reject_put
      security:
      - HTTPBearer: []
      parameters:
      - name: claim_id
        in: path
        required: true
        schema:
          type: integer
          title: Claim Id
      - name: notes
        in: query
        required: false
        schema:
          type: string
          title: Notes
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/claims/all/summary:
    get:
      tags:
      - Claims
      summary: All Claims Summary
      description: Get summary of all claims for admin dashboard.
      operationId: all_claims_summary_api_claims_all_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PolicyCreate:
      properties:
        worker_id:
          type: integer
          title: Worker Id
        plan_type:
          type: string
          title: Plan Type
          default: standard
      type: object
      required:
      - worker_id
      title: PolicyCreate
    PolicyResponse:
      properties:
        id:
          type: integer
          title: Id
        worker_id:
          type: integer
          title: Worker Id
        plan_type:
          type: string
          title: Plan Type
        weekly_premium:
          type: number
          title: Weekly Premium
        coverage_amount:
          type: number
          title: Coverage Amount
        status:
          type: string
          title: Status
        auto_renew:
          type: boolean
          title: Auto Renew
        start_date:
          type: string
          format: date-time
          title: Start Date
        end_date:
          type: string
          format: date-time
          title: End Date
        weeks_active:
          type: integer
          title: Weeks Active
        total_premiums_paid:
          type: number
          title: Total Premiums Paid
        total_claims_paid:
          type: number
          title: Total Claims Paid
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - worker_id
      - plan_type
      - weekly_premium
      - coverage_amount
      - status
      - auto_renew
      - start_date
      - end_date
      - weeks_active
      - total_premiums_paid
      - total_claims_paid
      - created_at
      title: PolicyResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer